wcd port modern dts hpx and dts sa+
Port of DTS Audio (both SA+ and Eagle) for newer versions of Android [than CAF supported]. SA+ is used as a fallback if Eagle cannot be started (check logs usually missing a *.so file somewhere). If you would like to use, please credit me in both authorship and relavent thread (if you declare my DTS port as a feature on your thread).

This one is for Michi but the defconfig breaks my brain and makes me wonder if we need to edit both WCD9330 and WCD9335.

Signed-off-by: Laster K. <officiallazerl0rd@gmail.com>